fix: widen the negotiate context accessors
One more instance of the class, found sweeping for the pattern rather than
listed in the issue. NegotiateContextDecoder.IsInvalid widens DataLength to
int and is sound, but Data recomputed 8+DataLength() in uint16: a context
declaring 65529 bytes wraps the upper bound to 1 and panics with ctx[8:1].

Reachable on the same pre-auth path as the negotiate response itself --
conn.go decodes the context list of an SMB311 response and calls Data() on
each entry -- and the transport accepts frames up to 16 MiB, so backing a
65529-byte context is no obstacle.

HashContextDataDecoder.Salt has the same wrap in 4+HashAlgorithmCount()*2 and
off+SaltLength. It has no callers today; widening it keeps a future one from
inheriting the panic.

+++ internal/smb2/smb2.go
@@ -149,7 +149,7 @@ func (ctx NegotiateContextDecoder) DataLength() uint16 {
 }
 
 func (ctx NegotiateContextDecoder) Data() []byte {
-	len := ctx.DataLength()
+	len := uint64(ctx.DataLength())
 	return ctx[8 : 8+len]
 }
 
@@ -191,8 +191,8 @@ func (h HashContextDataDecoder) HashAlgorithms() []uint16 {
 }
 
 func (h HashContextDataDecoder) Salt() []byte {
-	off := 4 + h.HashAlgorithmCount()*2
-	len := h.SaltLength()
+	off := 4 + uint64(h.HashAlgorithmCount())*2
+	len := uint64(h.SaltLength())
 	return h[off : off+len]
 }
